+++ internal/domain/entity/runtime_config.go
@@ -64,6 +64,7 @@ type RuntimeSearchShortcut struct {
 }
 
 type RuntimeOmniboxConfig struct {
+	Style             string
 	InitialBehavior   OmniboxInitialBehavior
 	MostVisitedDays   int
 	AutoOpenOnNewPane bool